+++ tests/fixtures.py
@@ -122,28 +122,12 @@ SCATACSEQ_BOTH_VERSIONS_VALID = {
     },
 }
 
-CONSTRAINTS_RESPONSE_GOOD = {
-    "code": 200,
-    "description": [
-        {
-            "code": 200,
-            "description": [
-                {"entity_type": "Sample", "sub_type": ["Suspension"], "sub_type_val": None},
-                {"entity_type": "Dataset", "sub_type": None, "sub_type_val": None},
-            ],
-            "name": "OK",
-        },
-        {
-            "code": 200,
-            "description": [
-                {"entity_type": "Sample", "sub_type": ["Block"], "sub_type_val": None},
-                {"entity_type": "Sample", "sub_type": ["Section"], "sub_type_val": None},
-                {"entity_type": "Sample", "sub_type": ["Suspension"], "sub_type_val": None},
-                {"entity_type": "Dataset", "sub_type": ["Light Sheet"], "sub_type_val": None},
-            ],
-            "name": "OK",
-        },
-    ],
-    "name": "OK",
-}
+CONSTRAINTS_RESPONSE_GOOD = b'{"code":200,"description":[{"code":200,"description":[{"entity_type":"Sample","sub_type":["Suspension"],"sub_type_val":null},{"entity_type":"Dataset","sub_type":null,"sub_type_val":null}],"name":"OK"},{"code":200,"description":[{"entity_type":"Sample","sub_type":["Block"],"sub_type_val":null},{"entity_type":"Sample","sub_type":["Section"],"sub_type_val":null},{"entity_type":"Sample","sub_type":["Suspension"],"sub_type_val":null},{"entity_type":"Dataset","sub_type":["Light Sheet"],"sub_type_val":null}],"name":"OK"}],"name":"OK"}'
+
 CONSTRAINTS_RESPONSE_BAD = b'{"code":400,"description":[{"code":200,"description":[{"entity_type":"Sample","sub_type":["Suspension"],"sub_type_val":null},{"entity_type":"Dataset","sub_type":null,"sub_type_val":null}],"name":"OK"},{"code":404,"description":[{"entity_type":"Sample","sub_type":["Block"],"sub_type_val":null},{"entity_type":"Sample","sub_type":["Section"],"sub_type_val":null},{"entity_type":"Sample","sub_type":["Suspension"],"sub_type_val":null},{"entity_type":"Dataset","sub_type":["Light Sheet"],"sub_type_val":null}],"name":"This `Sample` `block` cannot be associated with the provided `ancestors` due to entity constraints. Click the link to view valid entity types that can be `descendants`"}],"name":"Bad Request"}'
+
+SAMPLE_BLOCK_PARTIAL_RESPONSE_BAD = b'{"schema":{"name":"Sample Block"},"reporting":[{"errorType":"notStandardTerm","column":"processing_time_unit","row":0,"repairSuggestion":"minute","value":"min"},{"errorType":"notStandardTerm","column":"source_storage_duration_unit","row":0,"repairSuggestion":"minute","value":"min"}]}'
+
+SAMPLE_BLOCK_PARTIAL_RESPONSE_GOOD = b'{"schema":{"name":"Sample Block"},"reporting":[]}'
+
+SAMPLE_BLOCK_PARTIAL_ENTITY_API_RESPONSE = b'{"entity_type":"sample","sample_category":"block"}'